#10073: Merge rsqrt_bw to TTNN
