+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -820,8 +820,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.complex_abs_bw
 
-.. autofunction:: tt_lib.tensor.rsqrt_bw
-
 .. autofunction:: tt_lib.tensor.neg_bw
 
 .. autofunction:: tt_lib.tensor.lt_bw
