+++ docs/source/ttnn/ttnn/ttnn/rsqrt_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.rsqrt_bw:
+
+ttnn.rsqrt_bw
+#################
+
+.. autofunction:: ttnn.rsqrt_bw
