+++ tests/tt_eager/python_api_testing/unit_testing/backward_ops/test_backward_rsqrt.py
@@ -4,7 +4,7 @@
 
 import torch
 import pytest
-import tt_lib
+import ttnn
 from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import compare_pcc, data_gen_with_range
 
 
@@ -22,7 +22,7 @@ def test_bw_rsqrt(input_shapes, device):
 
     pyt_y = torch.rsqrt(in_data)
 
-    tt_output_tensor_on_device = tt_lib.tensor.rsqrt_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.rsqrt_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
